fix: serialize inference traces as float32

+++ src/openpi/policies/policy.py
@@ -115,11 +115,13 @@ class Policy(BasePolicy):
         if self._is_pytorch_model:
             outputs = jax.tree.map(lambda x: np.asarray(x[0, ...].detach().cpu()), outputs)
             if debug_trace is not None:
-                debug_trace = jax.tree.map(lambda x: np.asarray(x[0, ...].detach().cpu()), debug_trace)
+                debug_trace = jax.tree.map(
+                    lambda x: np.asarray(x[0, ...].detach().cpu(), dtype=np.float32), debug_trace
+                )
         else:
             outputs = jax.tree.map(lambda x: np.asarray(x[0, ...]), outputs)
             if debug_trace is not None:
-                debug_trace = jax.tree.map(lambda x: np.asarray(x[0, ...]), debug_trace)
+                debug_trace = jax.tree.map(lambda x: np.asarray(x[0, ...], dtype=np.float32), debug_trace)
 
         outputs = self._output_transform(outputs)
         outputs["policy_timing"] = {
